Explicitly set background color to white
When visiting the site from a mobile environment,
the header is almost unusable due to the header
background being very dark, and the font also being
fairly dark and difficult to see against the background.

Updating the app background to be white and set the
header text colours so header text is more readable.