Fix security console caching to reduce page load time
Update the useEffect for pre-loading security console
source images to be stateless without dependencies,
so that both desktop and mobile images are cached
and available for very fast page load times.

When adding a dependency on the useDeviceType custom
React hook to only load the specific image relevant
to the given environment (mobile vs desktop), the
page load time is much longer and we don't get the
benefit of instant image retrieval and caching.